requirements.txt
django==2.1.5
django-webpack-loader==0.6.0
djangorestframework==3.9.0
markdown==3.0.1
django-filter ==2.0.0